优化了账户控制层的代码，将学生和教师账户操作接口整合为用户操作接口。

+++ src/main/java/cn/edu/buaa/onlinejudge/mapper/TeacherMapper.java
@@ -11,7 +11,7 @@ public interface TeacherMapper {
      * @param teacherId - 教师ID
      * @return Teacher对象
      */
-    Teacher getTeacherById(int teacherId);
+    Teacher getTeacherById(long teacherId);
 
     /**
      * 根据邮箱（账号）获取教师对象
